data-engineering/requirements.txt
pandas==2.1.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
faker==22.0.0
numpy==1.26.2
streamlit==1.32.0
plotly==5.20.0
